Init SDK on app start, restart to change key/env
Brings SDK init logic more in line w/ iOS flappy birds app and is
closer to how real apps are supposed to use Snapyr SDK.

It's initialized on app start using stashed write key and env.
Splash screen allows changing these, but requires an app restart
to apply (so they're applied on the next launch).

This also ensures that the SDK is live regardless of which activity
was launched